Import renamed modules
