+++ src/modules/mealsList.js
@@ -1,4 +1,4 @@
-import { URL_API, getFoodList } from './listItems.js';
+import { URL_API, getFoodList } from './fetchMeals.js';
 // Using getFoodList function //
 const getdata = async () => {
   // create an empty array
